X/0 Game
An X/O game developed using the C programming language. The game takes place between the user and the computer. The user chooses the symbol he wants to be in the game. The number of moves in the game is limited and this limit is determined by the user before the game starts.